+++ lib/query/user_list/group_channel_member_list_query.g.dart
@@ -38,7 +38,7 @@ Map<String, dynamic> _$GroupChannelMemberListQueryToJson(
       'muted_member_filter':
           _$MutedMemberFilterEnumMap[instance.mutedMemberFilter],
       'member_state_filter':
-          _$MemberStateFilterEnumMap[instance.memberStateFilter],
+          memberStateFilterEnumForQuery(instance.memberStateFilter),
       'order': _$MemberListOrderEnumMap[instance.order],
     };